@@ -28,63 +22,63 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@angular/animations": "^8.2.14",
-    "@angular/common": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/forms": "^8.2.14",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
-    "core-js": "^2.6.3",
-    "ngx-bootstrap": "^3.2.0",
-    "rxjs": "^6.4.0",
-    "tslib": "^1.9.0",
-    "zone.js": "^0.9.1"
+    "@angular/animations": "9.1.7",
+    "@angular/common": "9.1.7",
+    "@angular/core": "9.1.7",
+    "@angular/forms": "9.1.7",
+    "@angular/platform-browser": "9.1.7",
+    "@angular/platform-browser-dynamic": "9.1.7",
+    "@angular/router": "9.1.7",
+    "ngx-bootstrap": "^5.6.1",
+    "rxjs": "~6.5.4",
+    "tslib": "^1.13.0",
+    "zone.js": "~0.10.2"
   },
   "peerDependencies": {
-    "rxjs": ">= 6.0.0"
+    "rxjs": ">= 5.5.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.23",
-    "@angular/cli": "^8.3.22",
-    "@angular/compiler": "^8.2.14",
-    "@angular/compiler-cli": "^8.2.14",
-    "@types/jasmine": "2.8.9",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "10.12.0",
-    "angular2-template-loader": "0.6.2",
-    "chalk": "2.4.1",
-    "codelyzer": "5.0.0",
-    "css-loader": "1.0.0",
-    "istanbul-instrumenter-loader": "3.0.1",
-    "jasmine-core": "3.2.1",
-    "karma": "^4.4.1",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^2.1.1",
+    "@angular-devkit/build-angular": "^0.901.7",
+    "@angular/cli": "~9.1.6",
+    "@angular/compiler": "9.1.7",
+    "@angular/compiler-cli": "9.1.7",
+    "@rollup/plugin-node-resolve": "^7.1.3",
+    "@types/jasmine": "^3.5.0",
+    "@types/jasminewd2": "^2.0.3",
+    "@types/node": "^14.0.1",
+    "angular2-template-loader": "^0.6.2",
+    "chalk": "^4.0.0",
+    "codelyzer": "^5.2.2",
+    "core-js": "~3.6.5",
+    "css-loader": "^3.5.3",
+    "dotenv": "^8.2.0",
+    "istanbul-instrumenter-loader": "^3.0.1",
+    "jasmine-core": "~3.5.0",
+    "karma": "^5.0.5",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-cli": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "^3.0.2",
     "karma-firefox-launcher": "^1.1.0",
-    "karma-jasmine": "1.1.2",
-    "karma-sourcemap-loader": "0.3.7",
-    "karma-spec-reporter": "0.0.32",
-    "karma-webpack": "4.0.0-rc.2",
+    "karma-jasmine": "^3.1.1",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-spec-reporter": "^0.0.32",
+    "karma-webpack": "^4.0.2",
     "npm-run-all": "^4.1.3",
-    "raw-loader": "0.5.1",
-    "reflect-metadata": "0.1.12",
-    "rollup": "0.66.6",
-    "rollup-plugin-license": "^2.0.0",
-    "rollup-plugin-node-resolve": "3.4.0",
-    "rollup-plugin-sourcemaps": "0.4.2",
-    "sass-loader": "7.1.0",
-    "shelljs": "0.8.2",
-    "source-map-loader": "0.2.4",
-    "to-string-loader": "1.1.5",
-    "ts-loader": "5.2.2",
-    "tsickle": "0.33.0",
+    "raw-loader": "^4.0.1",
+    "reflect-metadata": "^0.1.13",
+    "rollup": "^2.10.2",
+    "rollup-plugin-license": "^2.2.0",
+    "rollup-plugin-sourcemaps": "^0.6.2",
+    "sass-loader": "^8.0.2",
+    "shelljs": "^0.8.4",
+    "source-map-loader": "^0.2.4",
+    "ts-loader": "^7.0.4",
+    "tsickle": "0.38.1",
     "tslint": "^5.19.0",
     "tslint-angular": "^3.0.2",
-    "typescript": "^3.5.3",
-    "uglify-js": "3.4.9",
-    "webpack": "^4.41.5",
-    "webpack-filter-warnings-plugin": "1.2.0"
+    "typescript": "3.6.5",
+    "uglify-js": "^3.9.3",
+    "webpack": "^4.43.0",
+    "webpack-filter-warnings-plugin": "^1.2.1"
   }
 }
