@@ -71,11 +71,15 @@ generateItems(isPrepend: boolean) {
 }
 
 doPrepend() {
-  this.datasource.adapter.prepend(this.generateItems(true));
+  this.datasource.adapter.prepend({
+    items: this.generateItems(true)
+  });
 }
 
 doAppend() {
-  this.datasource.adapter.append(this.generateItems(false));
+  this.datasource.adapter.append({
+    items: this.generateItems(false)
+  });
 }`
   }, {
     active: true,
@@ -95,13 +99,15 @@ doAppend() {
   width: 150px;
   height: 250px;
   overflow-y: auto;
-  overflow-anchor: none;
 }
 .item {
   font-weight: bold;
   height: 25px;
 }`
   }];
+
+  appendCallSample = 'Adapter.append({ items })';
+  prependCallSample = 'Adapter.prepend({ items })';
 
   onInputChanged(target: HTMLInputElement) {
     const value = parseInt(target.value.trim(), 10);
@@ -123,11 +129,15 @@ doAppend() {
   }
 
   doPrepend() {
-    this.datasource.adapter.prepend(this.generateItems(true));
+    this.datasource.adapter.prepend({
+      items: this.generateItems(true)
+    });
   }
 
   doAppend() {
-    this.datasource.adapter.append(this.generateItems(false));
+    this.datasource.adapter.append({
+      items: this.generateItems(false)
+    });
   }
 
 }
