@@ -1,5 +1,6 @@
 import { ADAPTER_PROPS } from './props';
-import { ItemAdapter, AdapterPropType } from '../../interfaces/index';
+import { ItemAdapter } from '../../interfaces/index';
+import version from '../../../ui-scroll.version';
 
 let instanceCount = 0;
 
@@ -12,19 +13,22 @@ export class AdapterContext {
   id: number;
   mock: boolean;
 
-  constructor(mock?: boolean) {
+  constructor(mock: boolean) {
     const id = ++instanceCount;
 
     // props will be reassigned on Scroller instantiation
-    ADAPTER_PROPS(EMPTY_ITEM).forEach(({ name, value, type, permanent }) =>
-      Object.defineProperty(this, name, {
-        get: () => value,
-        configurable: !mock || permanent
-      })
-    );
+    ADAPTER_PROPS(EMPTY_ITEM)
+      .filter(({ permanent }) => !permanent)
+      .forEach(({ name, value, type }) =>
+        Object.defineProperty(this, name, {
+          get: () => value,
+          configurable: !mock
+        })
+      );
 
     // set up permanent props
     Object.defineProperty(this, 'id', { get: () => id, configurable: !mock });
-    Object.defineProperty(this, 'mock', { get: () => !!mock, configurable: !mock });
+    Object.defineProperty(this, 'mock', { get: () => mock, configurable: !mock });
+    Object.defineProperty(this, 'version', { get: () => version, configurable: !mock });
   }
 }
