@@ -1,12 +1,14 @@
-import { Direction } from '../interfaces/direction';
 import { Settings } from './settings';
+import { Direction } from '../interfaces/direction';
 
 export class Routines {
 
   readonly horizontal: boolean;
+  readonly window: boolean;
 
   constructor(settings: Settings) {
     this.horizontal = settings.horizontal;
+    this.window = settings.windowViewport;
   }
 
   checkElement(element: HTMLElement) {
@@ -16,20 +18,30 @@ export class Routines {
   }
 
   getScrollPosition(element: HTMLElement): number {
+    if (this.window) {
+      return window.pageYOffset;
+    }
     this.checkElement(element);
     return element[this.horizontal ? 'scrollLeft' : 'scrollTop'];
   }
 
   setScrollPosition(element: HTMLElement, value: number) {
-    this.checkElement(element);
     value = Math.max(0, value);
+    if (this.window) {
+      if (this.horizontal) {
+        window.scrollTo(value, window.scrollY);
+      } else {
+        window.scrollTo(window.scrollX, value);
+      }
+      return;
+    }
+    this.checkElement(element);
     element[this.horizontal ? 'scrollLeft' : 'scrollTop'] = value;
   }
 
-  getParams(element: HTMLElement): ClientRect {
+  getParams(element: HTMLElement, doNotBind?: boolean): ClientRect {
     this.checkElement(element);
-    if (element.tagName.toLowerCase() === 'body') {
-      element = element.parentElement as HTMLElement;
+    if (this.window && doNotBind) {
       return {
         'height': element.clientHeight,
         'width': element.clientWidth,
@@ -42,8 +54,8 @@ export class Routines {
     return element.getBoundingClientRect();
   }
 
-  getSize(element: HTMLElement): number {
-    return this.getParams(element)[this.horizontal ? 'width' : 'height'];
+  getSize(element: HTMLElement, doNotBind?: boolean): number {
+    return this.getParams(element, doNotBind)[this.horizontal ? 'width' : 'height'];
   }
 
   getSizeStyle(element: HTMLElement): number {
@@ -54,18 +66,14 @@ export class Routines {
 
   setSizeStyle(element: HTMLElement, value: number) {
     this.checkElement(element);
-    value = Math.max(0, value);
+    value = Math.max(0, Math.round(value));
     element.style[this.horizontal ? 'width' : 'height'] = `${value}px`;
   }
 
-  getRectEdge(params: ClientRect, direction: Direction, opposite?: boolean): number {
-    const forward = !opposite ? Direction.forward : Direction.backward;
-    return params[direction === forward ? (this.horizontal ? 'right' : 'bottom') : (this.horizontal ? 'left' : 'top')];
-  }
-
-  getEdge(element: HTMLElement, direction: Direction, opposite?: boolean): number {
-    const params = this.getParams(element);
-    return this.getRectEdge(params, direction, opposite);
+  getEdge(element: HTMLElement, direction: Direction, doNotBind?: boolean): number {
+    const params = this.getParams(element, doNotBind);
+    const isFwd = direction === Direction.forward;
+    return params[isFwd ? (this.horizontal ? 'right' : 'bottom') : (this.horizontal ? 'left' : 'top')];
   }
 
   getEdge2(element: HTMLElement, direction: Direction, relativeElement: HTMLElement, opposite: boolean): number {
@@ -81,7 +89,7 @@ export class Routines {
 
   getOffset(element: HTMLElement): number {
     this.checkElement(element);
-    return this.horizontal ? element.offsetLeft : element.offsetTop;
+    return (this.horizontal ? element.offsetLeft : element.offsetTop) || 0;
   }
 
   scrollTo(element: HTMLElement, argument?: boolean | ScrollIntoViewOptions) {
