@@ -1,7 +1,7 @@
 export class RenderModel {
   sizeBefore: number;
   sizeAfter: number;
-  fwdPaddingBefore: number;
+  positionBefore: number;
 
   get noSize(): boolean {
     return this.sizeBefore === this.sizeAfter;
@@ -14,6 +14,6 @@ export class RenderModel {
   reset() {
     this.sizeBefore = 0;
     this.sizeAfter = 0;
-    this.fwdPaddingBefore = 0;
+    this.positionBefore = 0;
   }
 }
