@@ -1,8 +1,9 @@
 import { Scroller } from '../../scroller';
-import { ADAPTER_METHODS } from '../../inputs/index';
 import { Process, ProcessStatus } from '../../interfaces/index';
 
 export default class Reload {
+
+  static process = Process.reload;
 
   static run(scroller: Scroller, reloadIndex: any) {
     const scrollPosition = scroller.viewport.scrollPosition;
