@@ -1,11 +1,13 @@
 import { Scroller } from '../../scroller';
 import { ADAPTER_METHODS, validate } from '../../inputs/index';
-import { Process, ProcessStatus, IDatasource, IDatasourceOptional } from '../../interfaces/index';
 import { Datasource } from '../../classes/datasource';
+import { Process, ProcessStatus, IDatasourceOptional } from '../../interfaces/index';
 
 export default class Reset {
 
-  static run(scroller: Scroller, params: IDatasourceOptional | null) {
+  static process = Process.reset;
+
+  static run(scroller: Scroller, params?: IDatasourceOptional) {
     const { datasource, buffer, viewport: { paddings } } = scroller;
 
     if (params) {
