@@ -1,12 +1,17 @@
 import { Scroller } from '../scroller';
-import { Process, ProcessStatus } from '../interfaces/index';
 import { Item } from '../classes/item';
+import { Process, ProcessStatus } from '../interfaces/index';
 
 export default class Render {
 
+  static process = Process.render;
+
   static run(scroller: Scroller) {
-    const { workflow, state: { clip, render } } = scroller;
+    const { workflow, state: { clip, render, scrollState }, viewport } = scroller;
     scroller.logger.stat('before new items render');
+    if (scrollState.positionBeforeAsync === null) {
+      scrollState.positionBeforeAsync = viewport.scrollPosition;
+    }
     scroller.innerLoopSubscriptions.push(
       scroller.bindData().subscribe(() => {
         const elts = scroller.viewport.element.querySelectorAll(`tr:not([data-sid])`);
@@ -32,24 +37,21 @@ export default class Render {
   }
 
   static processElements(scroller: Scroller): boolean {
-    const { state, state: { fetch, fetch: { items }, render }, viewport, buffer } = scroller;
-    const scrollBeforeRender = scroller.settings.windowViewport && fetch.isPrepend
-      ? scroller.viewport.scrollPosition
-      : null;
+    const { state: { fetch, render }, viewport, buffer, logger } = scroller;
+    render.positionBefore = viewport.scrollPosition;
     if (!fetch.isReplace) {
       render.sizeBefore = viewport.getScrollableSize();
-      render.fwdPaddingBefore = viewport.paddings.forward.size;
-      if (!items.reduce((acc, item) => acc && Render.processElement(scroller, item), true)) {
+      const success = fetch.items.reduce((acc, item) =>
+        acc && Render.processElement(scroller, item)
+      , true);
+      if (!success) {
         return false;
       }
     }
-    fetch.hasAverageItemSizeChanged = buffer.checkAverageSize();
+    buffer.checkAverageSize();
     render.sizeAfter = viewport.getScrollableSize();
-    if (!render.noSize && scrollBeforeRender !== null) {
-      Render.processWindowScrollBackJump(scroller, scrollBeforeRender);
-    }
-    scroller.logger.stat('after new items render');
-    scroller.logger.log(() => render.noSize ? 'viewport size has not been changed' : void 0);
+    logger.stat('after new items render');
+    logger.log(() => render.noSize ? 'viewport size has not been changed' : void 0);
     return true;
   }
 
@@ -65,29 +67,11 @@ export default class Render {
     item.element.classList.remove('temp');
     item.invisible = false;
     item.setSize();
-    buffer.cache.add(item);
+    buffer.cacheItem(item);
     if (item.$index < fetch.minIndex) {
       fetch.negativeSize += item.size;
     }
     return true;
-  }
-
-  static processWindowScrollBackJump(scroller: Scroller, scrollBeforeRender: number) {
-    const { state, state: { scrollState: { window } }, viewport } = scroller;
-    // if new items have been rendered in the area that is before current scroll position
-    // then this position will be updated silently in case of entire window scrollable
-    // so we need to remember the delta and to update scroll position manually right after it is changed silently
-    const inc = scrollBeforeRender >= viewport.paddings.backward.size ? 1 : -1;
-    const delta = inc * Math.abs(state.render.sizeAfter - state.render.sizeBefore);
-    const positionToUpdate = scrollBeforeRender - delta;
-    if (delta && positionToUpdate > 0) {
-      window.positionToUpdate = positionToUpdate;
-      window.delta = delta;
-      scroller.logger.log(() => {
-        const token = delta < 0 ? 'reduced' : 'increased';
-        return [`next scroll position (if ${positionToUpdate}) should be ${token} by`, Math.abs(delta)];
-      });
-    }
   }
 
 }
