@@ -3,11 +3,14 @@ import { Process, ProcessStatus } from '../interfaces/index';
 
 export default class Start {
 
+  static process = Process.start;
+
   static run(scroller: Scroller, process: Process, payload?: { process: Process }) {
-    const { state, state: { scrollState, fetch, clip, render }, adapter } = scroller;
+    const { state: { scrollState, fetch, clip, render }, adapter } = scroller;
     const processToPass = payload && payload.process || process;
 
     adapter.loopPending = true;
+    scrollState.positionBeforeAsync = null;
     if (!fetch.simulate) {
       fetch.reset();
     }
