@@ -68,7 +68,7 @@ const shouldClipAfterAppend = (config: TestBedConfig) => (misc: Misc) => (done: 
   let indexToAppend = -Infinity;
   const NEW_ITEMS_COUNT = 50;
   const { itemSize } = config.datasourceSettings;
-  const { buffer } = misc.scroller;
+  const { adapter, scroller: { buffer } } = misc;
   let firstIndex: number, lastIndex: number;
   const clipSettings = getClipArgument(config);
 
@@ -80,13 +80,13 @@ const shouldClipAfterAppend = (config: TestBedConfig) => (misc: Misc) => (done: 
         id: indexToAppend - x,
         text: `!item #${indexToAppend - x}`
       }));
-      misc.datasource.adapter.append(itemsToAppend);
+      adapter.append(itemsToAppend);
     } else if (cycles === 2) {
       firstIndex = buffer.firstIndex as number;
       lastIndex = buffer.lastIndex as number;
       expect(lastIndex).toEqual(indexToAppend + NEW_ITEMS_COUNT - 1);
       expect(misc.padding.backward.getSize()).toEqual(0);
-      misc.datasource.adapter.clip(clipSettings);
+      adapter.clip(clipSettings);
     } else if (firstIndex !== null) {
       // user clip requires additional reflow to remove DOM elements
       setTimeout(() => {
