@@ -7,6 +7,7 @@ export interface TemplateSettings {
   horizontal?: boolean;
   dynamicSize?: string | null;
   viewportPadding?: number;
+  headerHeight?: number;
 }
 
 const defaultTemplateSettings: TemplateSettings = {
@@ -25,17 +26,33 @@ export interface TemplateData {
   template: string;
 }
 
+const addPaddingFix = (padding?: number): string => {
+  // https://bugzilla.mozilla.org/show_bug.cgi?id=748518
+  if (!padding) {
+    return '';
+  }
+  const className = 'viewport-ff-padding';
+  const rules = `.${className}:after{content:'';height:${padding}px;display:block;}`;
+  const style = document.createElement('style');
+  style.textContent = rules;
+  document.head.append(style);
+  return ` ${className}`;
+};
+
 export const generateTemplate = (templateSettings?: TemplateSettings): TemplateData => {
   const settings = Object.assign({}, defaultTemplateSettings, templateSettings || {});
   const viewportClass = `${settings.noViewportClass ? '' :
-    'viewport' + (settings.horizontal ? '-horizontal' : '')}`;
+    'viewport' + (settings.horizontal ? '-horizontal' : '')}` + addPaddingFix(settings.viewportPadding);
   const viewportStyle = `${settings.viewportHeight ? 'height:' + settings.viewportHeight + 'px;' : ''}` +
     `${settings.viewportWidth ? 'width:' + settings.viewportWidth + 'px;' : ''}` +
     `${settings.viewportPadding ? 'padding:' + settings.viewportPadding + 'px;' : ''}`;
   const hasItemStyle = settings.dynamicSize || settings.itemHeight || settings.itemWidth;
+  const header = settings.headerHeight
+    ? `<div style="height:${settings.headerHeight}px;width:100%;background-color:yellow"></div>`
+    : '';
   return {
     settings,
-    template: `<div
+    template: `${header}<div
   class="${viewportClass}"
   style="${viewportStyle}"
 ><div
